Standardize `aws ecr get-login` commands
The first `aws ecr get-login` command in this page has the
`--region YOUR_REGION_HERE` parameter, with is not really necessary and
differs from the pattern used in other commands.

Also, the command output is multi-line and different from the other `aws
ecr get-login` command.

These changes are for keep things simpler and more standardized.