Fix region mismatch
The workshop is supposed to be region agnostic. But, in some pages, there
are lines that supposes the user is in the `us-east-1` region. In addition
to this, in the page `05-ContinuousDelivery`, there are lines that supposes
the user is in the `us-east-2` region. This may confuse users and lead them
into mistakes.

The changes are for make the workshop more region agnostic and to avoid
user mistakes.

Fixes: #16